+++ src/bluesearch/database/topic.py
@@ -42,6 +42,7 @@ class TopicInfo:
         self.articles_topics = {}
         self.journal_topics = {}
         self.source = None
+        self.other_metadata = {}
 
         self.bbs_version = bluesearch.version.__version__
 
@@ -82,6 +83,9 @@ class TopicInfo:
         if self.journal_topics:
             result["topics"]["journal"] = self.journal_topics
 
+        if self.other_metadata:
+            for key, value in self.other_metadata:
+                result["metadata"]["key"] = value
         return result
 
 
@@ -346,5 +350,61 @@ def get_topics_for_pmc_article(
     for mesh in journal_meshes:
         for descriptor in mesh["descriptor"]:
             journal_topics.append(descriptor["name"])
+
+    return journal_topics
+
+
+def extract_article_topics_for_pubmed_article(
+    xml_article: Element,
+) -> list[str] | None:
+    """Extract article topics of a PubMed article.
+
+    Parameters
+    ----------
+    xml_article
+        XML parse of an article for which to extract journal and article topics.
+
+    Returns
+    -------
+    article_topics : list[str] | None
+        Article topics extracted for the given article.
+    """
+    mesh_headings = xml_article.findall("./MedlineCitation/MeshHeadingList")
+    article_meshes = _parse_mesh_from_pubmed(mesh_headings)
+
+    article_topics = [
+        desc["name"] for mesh in article_meshes for desc in mesh["descriptor"]
+    ]
+
+    return article_topics
+
+
+def extract_journal_topics_for_pubmed_article(
+    xml_article: Element,
+) -> list[str] | None:
+    """Extract journal topics of a PubMed article.
+
+    Parameters
+    ----------
+    xml_article
+        XML parse of an article for which to extract journal and article topics.
+
+    Returns
+    -------
+    journal_topics : list[str] | None
+        Journal topics extracted for the given article.
+    """
+    # Journal topic
+    medline_ta = xml_article.find("./MedlineCitation/MedlineJournalInfo/MedlineTA")
+    if medline_ta is None or medline_ta.text is None:
+        return None
+
+    journal_meshes = request_mesh_from_nlm_ta(medline_ta.text)
+    if journal_meshes is None:
+        return None
+
+    journal_topics = [
+        desc["name"] for mesh in journal_meshes for desc in mesh["descriptor"]
+    ]
 
     return journal_topics