+++ tests/unit/entrypoint/database/test_topic_extract.py
@@ -165,3 +165,52 @@ def test_pmc_source(test_data_path, capsys, monkeypatch, tmp_path):
     assert exit_code == 0
     results = JSONL.load_jsonl(output_jsonl)
     assert len(results) == 2  # Length 2 because we append the file
+
+
+def test_pubmed_source(test_data_path, capsys, monkeypatch, tmp_path):
+    pmc_path = test_data_path / "pubmed_articles.xml"
+    output_jsonl = tmp_path / "test.jsonl"
+    journal_meshes = ["MeSH Journal 1", "MeSH Journal 2"]
+    article_meshes = ["MeSH Article 1", "MeSH Article 2"]
+
+    extract_article_topic_for_pubmed_mock = Mock(return_value=article_meshes)
+    monkeypatch.setattr(
+        "bluesearch.database.topic.extract_article_topics_for_pubmed_article",
+        extract_article_topic_for_pubmed_mock,
+    )
+
+    extract_journal_topic_for_pubmed_mock = Mock(return_value=journal_meshes)
+    monkeypatch.setattr(
+        "bluesearch.database.topic.extract_journal_topics_for_pubmed_article",
+        extract_journal_topic_for_pubmed_mock,
+    )
+
+    exit_code = topic_extract.run(
+        source="pubmed",
+        input_path=pmc_path,
+        output_file=output_jsonl,
+        match_filename=None,
+        recursive=False,
+        overwrite=False,
+        dry_run=False,
+    )
+    assert exit_code == 0
+    assert output_jsonl.exists()
+    results = JSONL.load_jsonl(output_jsonl)
+    assert extract_journal_topic_for_pubmed_mock.call_count == 2
+    assert extract_article_topic_for_pubmed_mock.call_count == 2
+
+    assert len(results) == 2
+    result = results[0]
+    assert result["source"] == "pubmed"
+    assert result["path"] == str(pmc_path)
+    assert isinstance(result["topics"], dict)
+    topics = result["topics"]
+    assert "article" in topics
+    assert "journal" in topics
+    assert isinstance(topics["journal"], dict)
+    assert isinstance(topics["article"], dict)
+    assert topics["journal"]["MeSH"] == journal_meshes
+    assert topics["article"]["MeSH"] == article_meshes
+    assert "metadata" in result
+    assert "element_in_file" in result["metadata"]